break out shared editor types
