+++ internal/editor/editor-funcs.go
@@ -6,26 +6,12 @@ import (
 	"strings"
 
 	"github.com/fieldse/gist-editor/internal/logger"
+	"github.com/fieldse/gist-editor/internal/shared"
 )
 
 // TextSelection represents the position and content of the editor's current text selection
-type TextSelection struct {
-	Position Position // cursor position, using rows and columns
-	Content  string
-}
-
-// TODO -- this might be useful for clarifying the exact position of the character selection
-type AbsoluteCharacterRange struct {
-	Start int
-	End   int
-}
-
-// Position represents a cursor position in a multi-row selection.
-// Both row and column values start from 1
-type Position struct {
-	Col int
-	Row int
-}
+type TextSelection = shared.TextSelection
+type Position = shared.Position
 
 // EditorFunctions is the set of Markdown syntax operations that can be performed
 // on the editor text content
